requirements.txt
boto3
botocore
certifi
chardet
click
dataclasses
docutils
filelock
jmespath
joblib
nltk
numpy
python-dateutil
regex
requests
s3transfer
sacremoses
sentencepiece
six
tensorflow
tensorflow-hub
tokenizers
torch
tqdm
transformers
urllib3
textattack
python-dotenv
wandb